+++ COSC363A1/base_object.h
@@ -3,5 +3,5 @@ class SceneObject{
 
 public:
 	virtual void Render();
-	virtual void Process(double dt);
+	virtual void Process(float dt);
 };
