+++ COSC363A1/factory.h
@@ -26,6 +26,6 @@ public:
 	Factory();
 	~Factory();
 	virtual void Render();
-	virtual void Process(double dt);
+	virtual void Process(float dt);
 };
